renames monster_module to monster_manager
