+++ monster_manager.py
